feat: criando arquivo filter-users list para apartar a lógica do appcomponent
