+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { IUser } from './interfaces/user/user.interface';
 import { UsersList } from './data/users-list';
 import { IFilterOptions } from './interfaces/filter_options.interface';
-import { isWithinInterval } from 'date-fns';
+import { filterUsersList } from './utils/filter-users-list';
 
 @Component({
   selector: 'app-root',
@@ -31,57 +31,6 @@ export class AppComponent implements OnInit {
   public filterSelected(filterOptions: IFilterOptions) {
     console.log(filterOptions);
 
-    this.userListFiltered = this.filterUsersList(filterOptions, this.usersList)
+    this.userListFiltered = filterUsersList(filterOptions, this.usersList)
   }
-
-  public filterUsersList(filterOptions: IFilterOptions, usersList: IUser[]): IUser[] {
-    let filteredList: IUser[] = [];
-
-    filteredList = this.filterUsersListByName(filterOptions.name, usersList);
-    filteredList = this.filterUsersListByStatus(filterOptions.status, filteredList);
-    filteredList = this.filterUsersListByDate(filterOptions.startDate, filterOptions.endDate, filteredList);
-
-    return filteredList;
-  }
-
-  public filterUsersListByDate(startDate: Date | undefined, endDate: Date | undefined, usersList: IUser[]): IUser[] {
-    const DATE_NOT_SELECTED = startDate === undefined || endDate === undefined;
-
-    if (DATE_NOT_SELECTED) {
-      return usersList;
-    }
-
-    const filteredList = usersList.filter((user) => isWithinInterval(new Date(user.dataCadastro), {
-      start: startDate,
-      end: endDate
-    }));
-
-    return filteredList;
-  }
-
-  public filterUsersListByStatus(status: boolean | undefined, usersList: IUser[]): IUser[] {
-    const STATUS_NOT_SELECTED = status === undefined;
-
-    if (STATUS_NOT_SELECTED) {
-      return usersList;
-    }
-
-    const filteredList = usersList.filter((user) => user.ativo === status);
-
-    return filteredList;
-  }
-
-  public filterUsersListByName(name: string | undefined, usersList: IUser[]): IUser[] {
-    const NAME_NOT_TYPED = name === undefined;
-
-    if (NAME_NOT_TYPED) {
-      return usersList;
-    }
-
-    const filteredList = usersList.filter((user) => user.nome.toLocaleLowerCase().includes(name.toLocaleLowerCase()));
-
-    return filteredList;
-  }
-
-
 }
